repr {repr}	R Documentation

Dynamic representation

Description

Specify an object and a format to represent it in. Will stop() if no such format is known.

Usage


repr(obj, format = "text", ...)

Arguments

obj
The object to create a representation for

format
The representation format. repr_<format> is then called. (default: Call repr_text)

...
delegated to the specific repr_<format> function

Value

A character or raw vector of that format or NULL if none is defined. Only the 'text' format is defined for everything (via print())

See Also

repr_text, repr-generics

[Package repr version 0.19.2 Index]