repr-options {repr}	R Documentation

repr options

Description

These options are used to control the behavior of repr when not calling it directly. Use options(repr.* = ...) and getOption('repr.*') to set and get them, respectively.

Usage


repr_option_defaults

Format

An object of class list of length 14.

Details

Once this package is loaded, all options are set to defaults which weren’t set beforehand.

Setting all options set to NULL are reset to defaults when reloading the package (or calling repr:::.onload()).

Options

repr.plot.*

Those are for representations of recordedplot instances:

repr.plot.width

Plotting area width in inches (default: 7)

repr.plot.height

Plotting area height in inches (default: 7)

repr.plot.pointsize

Text height in pt (default: 12)

repr.plot.bg

Background color (default: white)

repr.plot.antialias

Which kind of antialiasing to use for for lines and text? 'gray', 'subpixel' or 'none'? (default: gray)

repr.plot.res

PPI for rasterization (default: 120)

repr.plot.quality

Quality of JPEG format in % (default: 90)

repr.plot.family

Vector font family. 'sans', 'serif', 'mono' or a specific one (default: sans)

repr.vector.quote

Output quotation marks for character vectors? (default: TRUE)

repr.matrix.max.rows

How many rows to display at max. Will insert a row with vertical ellipses to show elision. (default: 60)

repr.matrix.max.cols

How many cols to display at max. Will insert a column with horizontal ellipses to show elision. (default: 20)

repr.matrix.latex.colspec

How to layout LaTeX tables when representing matrices or data.frames. List of row.head, other col, and end strings. end mainly exists for when you want a vertical line there (default: 'r|', 'l', and ”)

repr.function.highlight

Use the highr package to insert highlighting instructions into the code? Needs that package to be installed. (default: FALSE)

repr.html.deduplicate

Use the html_dependencies manager to only include dependencies once? This can greatly reduce notebook size, but fails if e.g. iframes are used (default: FALSE)

[Package repr version 0.19.2 Index]