repr_text {repr}	R Documentation

Text representation

Description

The only representation defined per default for everthing (via print())

Usage


repr_text(obj, ...)

## Default S3 method:
repr_text(obj, ...)

Arguments

obj
The object to print and then return the output

...
ignored

See Also

repr-generics for other generics

[Package repr version 0.19.2 Index]